migrate TokenVocabulary to u16 to prep for tokenization
Char-level is philosophically elegant, but we're so bottlenecked on
compute as a 1 CPU project, that I'm going to need to tokenize to see
how far this can go.

Claude Code changed the types here, but this is trivial enough that
I'm not going to neurotically give it a Co-Authored-By credit like I
did the other day; it really is a tool-level rather than an
author-level contribution.

+++ src/core/embedding.rs
@@ -10,8 +10,8 @@ use super::{Origin, Parameterized, Tensor, TensorBuilder};
 
 pub struct TokenVocabulary {
     // TODO: getter methods (that don't return Option) instead of pub HashMap
-    pub token_to_id: HashMap<char, u8>,
-    pub id_to_token: HashMap<u8, char>,
+    pub token_to_id: HashMap<char, u16>,
+    pub id_to_token: HashMap<u16, char>,
 }
 
 impl TokenVocabulary {
@@ -19,8 +19,8 @@ impl TokenVocabulary {
         let mut token_to_id = HashMap::new();
         let mut id_to_token = HashMap::new();
         for (i, token) in tokens.iter().enumerate() {
-            id_to_token.insert(i as u8, *token);
-            token_to_id.insert(*token, i as u8);
+            id_to_token.insert(i as u16, *token);
+            token_to_id.insert(*token, i as u16);
         }
         TokenVocabulary {
             token_to_id,
@@ -36,7 +36,7 @@ impl TokenVocabulary {
         let mut token_ids = Vec::new();
         for c in text.chars() {
             match self.token_to_id.get(&c) {
-                Some(id) => token_ids.push(*id as u8 as f32),
+                Some(id) => token_ids.push(*id as f32),
                 None => {}
             }
         }
